Hra rozdelena na 4 moduly + modul s testy
