+++ domaci_ukoly/piskvorky/hra.py
@@ -1,0 +1,6 @@
+# 1-D piskvorky, hraji se na radku s dvaceti policky. 
+# Hrac clovek ma krizky, pocitac ma kolecka. Kdo da tri sve symboly vedle sebe, vyhral.
+
+from piskvorky import piskvorky1d
+
+piskvorky1d()
